Login and station page added
I've created a general function to create buttons in rows. The number of
buttons are controlled by a parameter, as is the size of the button.
This seems to properly scale to some not unreasonable number of buttons.
Additional work needs to be done to fully generalize this so it will
work for any data.

The login and station views are generally on the way to completion. I
have some temporary fake data in place so I can generate the necessary
views, but I expect the actual data structure to change, though it
shouldn't be difficult to fix that when the time comes.

I'll get the "Currently Playing" view working next, after completing the
necessary generalization work to fix the button generation function.